:rotating_light: resolve kt lint failure
